back/requirements.txt
# Servidor y Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Base de Datos
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Seguridad y Autenticación
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
email-validator==2.1.0.post1

# Validación y Entorno
pydantic[email]==2.5.3
python-dotenv==1.0.1
pymupdf==1.24.1
# Calidad de Código (Linter)
ruff==0.1.14

#Ia
httpx==0.27.0